Whole-string match reqs in comment_requirement (#55192)
